+++ spec/dummy/app/components/previews/shadcn/calendar_preview.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Shadcn
+  class CalendarPreview < ViewComponent::Preview
+    def default
+      render(Shadcn::Calendar.new(month: Date.new(2026, 8, 1), selected: Date.new(2026, 8, 27)))
+    end
+
+    def plain
+      render(Shadcn::Calendar.new(month: Date.new(2026, 8, 1)))
+    end
+  end
+end