+++ spec/dummy/app/controllers/pages_controller.rb
@@ -21,6 +21,11 @@ class PagesController < ApplicationController
 
   def sidebar; end
 
+  def calendar
+    parsed = Date.parse("#{params[:month]}-01") rescue Date.current.beginning_of_month
+    @month = parsed
+  end
+
   def carousel; end
 
   def form